+++ pkg/traces/automaticloggingprocessor/factory.go
@@ -110,7 +110,7 @@ func NewFactory() component.ProcessorFactory {
 	return component.NewProcessorFactory(
 		TypeStr,
 		createDefaultConfig,
-		component.WithTracesProcessor(createTraceProcessor),
+		component.WithTracesProcessor(createTraceProcessor, component.StabilityLevelUndefined),
 	)
 }